use <nixpkgs> of nixos-unstable & clean sources
[sources] init (1), update (20), remove (3)

gitignore: 2020-07-17 -> 2021-01-26
niv: 2021-01-14 -> 2021-05-11
nix-gl: 2020-10-31 -> 2021-04-28
nixpkgs: 2021-01-18 -> 2021-05-15
nixpkgs-unstable: remove
nixpkgs-unstable-bb010g: remove
nur: 2021-01-19 -> 2021-05-15
vim-ale: 2021-01-19 -> 2021-05-23
vim-caw: 2020-12-01 -> 2021-04-15
vim-context-filetype: 2020-09-17 -> 2021-05-13
vim-diffchar: 2020-05-01 -> 2021-04-16
vim-eunuch: init at 2021-04-05
vim-gina: 2020-10-07 -> 2021-04-28
vim-polyglot: 2021-01-14 -> 2021-04-14
vim-precious: 2019-08-23 -> 2021-05-15
vim-recover: 2020-04-20 -> 2021-05-18
vim-repeat: 2019-11-13 -> 2021-01-25
vim-scriptease: 2020-01-05 -> 2021-03-25
vim-suckless: remove
vim-suda: 2020-09-08 -> 2021-02-20
vim-table-mode: 2020-10-13 -> 2021-05-05
vim-undotree: 2020-11-10 -> 2021-05-16
zsh-fast-syntax-highlighting: 2020-11-26 -> 2021-05-14
zsh-nix-shell: 2020-09-24 -> 2021-04-30

+++ update-firefox-profiles.sh
@@ -23,12 +23,14 @@ jq_fxpath() {
 typeset -l profile file
 for profile ( ~/.mozilla/firefox/*.default(/) ) {
   for file ( extensions.json ) {
+    if [[ ! -e $profile/$file ]]; then continue; fi
     jq_fxpath -c $profile/$file > $profile/$file.new
     if [[ "$(jq -c '.' $profile/$file | wc -c)" = "$(jq -c '.' $profile/$file.new | wc -c)" ]] {
       mv $profile/$file{.new,}
     }
   }
   for file ( addonStartup.json.lz4 ) {
+    if [[ ! -e $profile/$file ]]; then continue; fi
     mozlz4-tool -d $profile/$file \
     | jq_fxpath -c \
     | mozlz4-tool -c /dev/fd/0 > $profile/$file.new
